Commentaries update!!!

+++ src/struct_managers/env_manager.c
@@ -3,15 +3,21 @@
 /*                                                        :::      ::::::::   */
 /*   env_manager.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mcrenn <mcrenn@student.42angouleme.fr>     +#+  +:+       +#+        */
+/*   By: ansimonn <ansimonn@student.42angouleme.f>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/20 10:45:19 by ansimonn          #+#    #+#             */
-/*   Updated: 2026/06/01 10:48:05 by mcrenn           ###   ########.fr       */
+/*   Updated: 2026/06/01 14:11:36 by ansimonn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
+/**
+ * @brief Counts the number of nodes in the 'env' list
+ *
+ * @param env List of environmental variables
+ * @return Number of nodes in env, or 0 if env is NULL
+ */
 int	get_env_size(const t_env *env)
 {
 	int	size;
@@ -25,9 +31,16 @@ int	get_env_size(const t_env *env)
 	return (size);
 }
 
+/**
+ * @brief Searches for a pointer to the value of the 'name' variable in env
+ *
+ * @param env List of environmental variables
+ * @param name Name of the variable we want to get
+ * @return Pointer to the value of the variable 'name', or NULL if not found
+ */
 char	**get_env(t_env *env, const char *name)
 {
-	int	size;
+	size_t	size;
 
 	if (!name || !env)
 		return (NULL);
@@ -41,6 +54,13 @@ char	**get_env(t_env *env, const char *name)
 	return (NULL);
 }
 
+/**
+ * @brief Sets the value of the 'var_name' variable to 'new_val', if existing
+ *
+ * @param var_name Name of the variable we want to change
+ * @param new_val New value to set to the 'var_name' variable in env
+ * @param env List of environmental variables
+ */
 void	set_env(const char *var_name, char *new_val, t_env *env)
 {
 	char	**var_val;
@@ -56,6 +76,14 @@ void	set_env(const char *var_name, char *new_val, t_env *env)
 	}
 }
 
+/**
+ * @brief Adds a new variable at the end of 'env'
+ *
+ * @param env List of environmental variables
+ * @param name Name of the variable to add
+ * @param value Value of the variable to add
+ * @return 0 on success, 1 if an error occurred
+ */
 int	add_env(t_env *env, char *name, char *value)
 {
 	char	**get;
@@ -78,6 +106,11 @@ int	add_env(t_env *env, char *name, char *value)
 	return (0);
 }
 
+/**
+ * @brief Increases the SHLVL variable in env by 1, if existing
+ *
+ * @param env List of environmental variables
+ */
 void	up_shlvl(t_env *env)
 {
 	int		conv;
